+++ packages/pyre-ast.0.1.3/lib/taglessFinal/binding.c
@@ -33,6 +33,9 @@ static value PyUnicode_to_ocaml_string(PyObject *object) {
   // Explicitly get the size to avoid issues with embedded null bytes.
   Py_ssize_t size;
   const char *data = PyUnicode_AsUTF8AndSize(object, &size);
+  if (data == NULL) {
+    caml_failwith("UTF8 casting failed");
+  }
   result = caml_alloc_initialized_string(size, data);
 
   CAMLreturn(result);
@@ -50,6 +53,26 @@ static value PyBytes_to_ocaml_string(PyObject *object) {
     caml_failwith("Unable to read byte string literal");
   }
   result = caml_alloc_initialized_string(size, buffer);
+
+  CAMLreturn(result);
+}
+
+// Mostly similar to PyUnicode_to_ocaml_string, except the string gets encoded
+// instead of casted to UTF8. This function can handle more inputs than
+// PyUnicdoe_to_ocaml_string, but it cost one more allocation. Avoid calling
+// this if the given object is known to be UTF8-encoded to begin with.
+static value PyUnicode_to_encoded_ocaml_string(PyObject *object) {
+  CAMLparam0();
+  CAMLlocal1(result);
+
+  PyObject *bytes_object =
+      PyUnicode_AsEncodedString(object, "utf-8", "backslashreplace");
+  if (bytes_object == NULL) {
+    caml_failwith("UTF8 encoding failed");
+  }
+  // TODO: If this call raises, we would leak `bytes_object`
+  result = PyBytes_to_ocaml_string(bytes_object);
+  Py_DECREF(bytes_object);
 
   CAMLreturn(result);
 }
@@ -369,7 +392,7 @@ CAMLprim value visit_constant(value visitor_value, PyObject *object) {
     }
   } else if (PyUnicode_Check(object)) {
     result = caml_callback(CONSTANT_STRING(visitor_value),
-                           PyUnicode_to_ocaml_string(object));
+                           PyUnicode_to_encoded_ocaml_string(object));
   } else if (PyFloat_Check(object)) {
     result = caml_callback(CONSTANT_FLOAT(visitor_value),
                            caml_copy_double(PyFloat_AsDouble(object)));
@@ -2250,8 +2273,10 @@ static struct RawModule RawModule_val(value v) {
   return result;
 }
 
-CAMLprim value cpython_parse_module(value filename_value, value input_value) {
-  CAMLparam2(filename_value, input_value);
+CAMLprim value cpython_parse_module(value filename_value,
+                                    value type_comment_value,
+                                    value input_value) {
+  CAMLparam3(filename_value, type_comment_value, input_value);
   CAMLlocal1(result);
 
   PyObject *filename = PyUnicode_FromString(String_val(filename_value));
@@ -2269,7 +2294,9 @@ CAMLprim value cpython_parse_module(value filename_value, value input_value) {
   }
 
   PyCompilerFlags flags = _PyCompilerFlags_INIT;
-  flags.cf_flags |= PyCF_TYPE_COMMENTS;
+  if (Bool_val(type_comment_value)) {
+    flags.cf_flags |= PyCF_TYPE_COMMENTS;
+  }
   mod_ty ast = _PyParser_ASTFromString(String_val(input_value), filename,
                                        Py_file_input, &flags, arena);
 
@@ -2304,11 +2331,9 @@ CAMLprim value cpython_parse_expression(value input_value) {
                         DEFAULT_SYNTAX_ERROR_LINE, DEFAULT_SYNTAX_ERROR_COLUMN);
   }
 
-  PyCompilerFlags flags = _PyCompilerFlags_INIT;
-  flags.cf_flags |= PyCF_TYPE_COMMENTS;
   mod_ty ast = _PyParser_ASTFromString(String_val(input_value),
                                        _PyUnicode_FromId(&Dummy_filename),
-                                       Py_eval_input, &flags, arena);
+                                       Py_eval_input, NULL, arena);
   if (ast == NULL) {
     _PyArena_Free(arena);
     raise_parsing_error_from_last_python_exception();
@@ -2339,11 +2364,9 @@ CAMLprim value cpython_parse_function_type(value input_value) {
                         DEFAULT_SYNTAX_ERROR_LINE, DEFAULT_SYNTAX_ERROR_COLUMN);
   }
 
-  PyCompilerFlags flags = _PyCompilerFlags_INIT;
-  flags.cf_flags |= PyCF_TYPE_COMMENTS;
   mod_ty ast = _PyParser_ASTFromString(String_val(input_value),
                                        _PyUnicode_FromId(&Dummy_filename),
-                                       Py_func_type_input, &flags, arena);
+                                       Py_func_type_input, NULL, arena);
   if (ast == NULL) {
     _PyArena_Free(arena);
     raise_parsing_error_from_last_python_exception();